rename test demo to events demo
